Model refactor
