+++ src/models/delete.js
@@ -1,0 +1,15 @@
+// drop todos
+const deleteTask = (task, element) => {
+    const taskName = task.children[0].children[1].value;
+    const tasks = JSON.parse(localStorage.getItem('tasks'));
+    const taskIndex = tasks.findIndex((x) => x.description === taskName);
+    tasks.splice(taskIndex, 1);
+    tasks.forEach((item, ind) => {
+      item.index = ind + 1;
+    });
+    localStorage.setItem('tasks', JSON.stringify(tasks));
+    element.parentElement.parentElement.remove();
+  };
+
+  export default deleteTask;
+
